Make the two sequence-number spaces distinct types
Both were uint64, which is why mixing them compiled, and that cost three
bugs, every one silent:

  - Adoption stored one number and used it for both. The new server began
    its client log at the shim's position while a reconnecting client
    resumed from a post-rewrite position, and Subscribe clamps a position
    past the end to the end, so the difference was skipped without a word.
    Nine bytes per prompt marker, 27 across three commands: enough to eat
    the front of an escape sequence and leave the rest as literal text.
  - `cm read --since-commands` anchored on pre-rewrite boundaries and read
    them out of the post-rewrite log, drifting all session.
  - modelSeq needed a comment to say which space it was in.

internal/seq names them, seq.Shim and seq.Log, and seqlog is generic over
the space so a log is tied to one. The documented mistake is now a compile
error: `recent.Subscribe(lastSeq)` gives "cannot use s.lastSeq (variable
of uint64 type seq.Shim) as seq.Log value", while the same call with
modelSeq builds. Both were checked.

The conversions that remain are the honest crossings, and each says why:
the protobuf wire, the sqlite row, and one deliberate fallback in adoption
for a database written before client_seq existed. Two tests that compare
across the spaces on purpose now say so explicitly, which is the point
rather than a cost: resumespace_test asserts the two drift apart, and it
should be obvious that it means to.

No behaviour change. The full suite is green including the 180s e2e run,
which is what covers the paths a type change cannot.

+++ internal/server/resumespace_test.go
@@ -60,7 +60,9 @@ func TestResumePointsUseSeparateNumberingSpaces(t *testing.T) {
 	if shimSeq == 0 || clientSeq == 0 {
 		t.Fatalf("resumePoints() = (%d, %d), want both advanced past zero", shimSeq, clientSeq)
 	}
-	if clientSeq <= shimSeq {
+	// Compared across the two spaces on purpose, which is what this test is for, so the conversion is
+	// explicit. Everywhere else the types make this a compile error; see internal/seq.
+	if uint64(clientSeq) <= uint64(shimSeq) {
 		t.Errorf("resumePoints() = (shim %d, client %d), want the client position ahead.\n"+
 			"The rewrite appends \";redraw=0\" to each prompt marker, so the bytes clients received are "+
 			"longer than the bytes the shim sent. Equal values mean one number is being used for both "+
@@ -107,7 +109,7 @@ func TestAdoptedSessionServesAResumingClient(t *testing.T) {
 	sub.Close()
 
 	shimSeq, clientSeq := first.resumePoints()
-	if clientSeq == shimSeq {
+	if uint64(clientSeq) == uint64(shimSeq) {
 		t.Fatalf("the two positions did not diverge (both %d), so this test would pass either way",
 			shimSeq)
 	}
@@ -164,7 +166,7 @@ func TestAdoptFallsBackWhenClientSeqIsUnset(t *testing.T) {
 	}
 	defer sess.Close()
 
-	if got := sess.recent.Next(); got != rec.LastSeq {
+	if got := sess.recent.Next(); uint64(got) != uint64(rec.LastSeq) {
 		t.Errorf("adopted log begins at %d, want the LastSeq fallback %d.\n"+
 			"Zero in client_seq means the record predates the column. Taking it literally would start "+
 			"the log at 0, so every client position would look like the distant future.",
